% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exmo.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by InitBioLink.}
}
\description{
End the module analysis
}
\examples{
   res <- InitMo()
   res <- LoadMO(PID=res$PID, UseExample = "example#1", DataPath = NULL, VocaPath = NULL)
   res2 <- MulOmicsCros(PID=res$PID, OutPath ="default", OmicGroups = "immunome,metabolome,proteome",
   VarsY = "Y1", VarsC = "all.c", TuneMethod = "random_search", TuneNum = 5, RsmpMethod = "cv", Folds = 5,
   Ratio = 0.67, Repeats = 5, VarsImpThr = 0.85, SG_Lrns ="lasso,enet,rf,xgboost")
   res3 <- VizMulOmicCros(PID=res$PID, OutPath = "default", VarsY = "Y1", NodeNum=100, EdgeThr= 0.45,
   Layout = "force-directed", Brightness = "light", Palette = 'default1')
   FuncExit(PID = res$PID)
}
\author{
Guohuan Zhang, Yuting Wang, Bin Wang (corresponding author)
}
